% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_linkPages_fn.R
\name{get_linkPages}
\alias{get_linkPages}
\title{Makes connection between child and parent pages using a link
and grabs page numbers of SDTM variables from aCRF file}
\usage{
get_linkPages(dsin = NULL, dset_defOrigin = NULL)
}
\arguments{
\item{dsin}{preprocessed input dataset}

\item{dset_defOrigin}{modified Define Origin file}
}
\description{
Makes connection between child and parent pages using a link
and grabs page numbers of SDTM variables from aCRF file
}
