% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aCRFExtractor_mainFn.R
\name{run_aCRFExtractor}
\alias{run_aCRFExtractor}
\title{Extracts page numbers of SDTM variables available in manually created aCRF file. The output is used for creation of Define XML file.
The input files (e.g., studyid, aCRF file, and Define Origin file) are taken from a user via console}
\usage{
run_aCRFExtractor()
}
\description{
Extracts page numbers of SDTM variables available in manually created aCRF file. The output is used for creation of Define XML file.
The input files (e.g., studyid, aCRF file, and Define Origin file) are taken from a user via console
}
\author{
Hailemichael M. Worku (aka, Haile). Email: <hailemichael.x.worku@gsk.com>
}
